Update longitud de circunferencia.vb

+++ longitud de circunferencia.vb
@@ -3,7 +3,7 @@ Imports System
 Module Program
 	Sub Main()
 		Console.Title = "Rony Rodriguez"
-		Dim radio, pi, Longitud As Integer
+		Dim radio, pi, Longitud As Decimal
 		pi = 3.1416
 		Console.WriteLine("Ingrese el radio del dado: ")
 		radio = Console.ReadLine()
@@ -11,7 +11,7 @@ Module Program
 		Console.WriteLine("La longitud del radio es: " & Calcular(pi, radio))
 
 	End Sub
-	Function Calcular(pi, radio) As Integer
+	Function Calcular(pi, radio) As Decimal
 		Return (pi * (radio * radio))
 	End Function
 End Module
